requirements.txt
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
pandas==2.0.3
pillow==10.0.0
matplotlib==3.7.2
scikit-learn==1.3.0
opencv-python==4.8.0.76
